dmshow.py: Print mapped_device->flags
Currently dmshow program prints the mapped_device->flags in hex format
only. It would require manual decoding to verify if the device is in
suspended state.

This patch updates dmshow program to print device flags if the device
is suspended, frozen or being deleted. This would help to verify if the
processes were stuck in UN state, waiting on IO requests on any of the
affected dm-X device maps.