Add a check for SDEV_TRANSPORT_OFFLINE state
The SDEV_TRANSPORT_OFFLINE device state was recently added to RHEL 6.x
kernels. But this state was not added in 'scsi_device_state' enum,
instead it was defined separately:

include/scsi/scsi_device.h
        #define SDEV_TRANSPORT_OFFLINE  (SDEV_CREATED_BLOCK + 1)

As the above state is not defined in 'scsi_device_state' enum, it could
not be processed by enum_sdev_state.getnam method. This additional check
takes care of a special case were scsi_device state is
SDEV_TRANSPORT_OFFLINE, but it is not defined in scsi_device_state enum.