dmshow: merge in upstream with downstream dmshow
For a long while I have been updating the downstream dmshow based
on testing done in the retrace hooks framework.  I've also added
features for various target types.

As the downstream was 1k lines longer than the upstream, instead of
merging downstream into upstream, i did the opposite.  I took all
the commits that have been added to upstream since upload and ensured
they are all present in the downstream version.

The only conflicts I had were with 29c91afb6cdd and 1d17ae90d83e.
This was because I'd patched the same issue a different way in the
downstream version.  When this happened, I tried my best to pull in
the features of both patches to make sure all corners were covered.

Signed-off-by: John Pittman <jpittman@redhat.com>