dmshow: supress failures from scope set
Due to a bug (https://bugzilla.redhat.com/show_bug.cgi?id=1811208),
sometimes we can get a error message when trying to set the scope
in crash.  Tell pykdump to ignore this error message.  The scope
sets are there for legacy purposes; we've seen in the past commands
fail without the scope being set manually.

Signed-off-by: John Pittman <jpittman@redhat.com>