scsishow.py: Add a check to verify FC rport status
With this patch the 'scsishow --check' command now verifies the status
of scsi_target. If the target status is not STARGET_RUNNING, and it is
an FC target, then program will also try to verify the status of
underlying fc rport.

In case any FC rport is not in FC_PORTSTATE_ONLINE state, then a warning
would be logged along with rport's WWPN.

This patch also improves the formatting of summary reported by --check
option.