Use separate functions to check for scsi host, device, cmd errors
The current copy of run_scsi_checks() function in scsishow.py runs
the automated checks for scsi host, device, command and target
errors. This function is now over 173 lines long. With this patch it is
split in 3 separate sub functions to perform checks for host, device,
commands and target issues:
- run_host_checks()
- run_sdev_cmd_checks()
- run_target_checks()