Add a check to verify if the scsi_device has corresponding gendisk struct
The '-q' option of scsishow.py did not check if the device has
corresponding gendisk struct or not. And fetching the gendisk struct for
devices with no valid pointer for gendisk was crashing the program.

This patch fixes the handling of scsi devices which do not have
corresponding gendisk struct.