scsishow: Print FC rports connected to QLogic FastLinQ HBAs
The existing code in print_fcrports() is able to process the details of
FC rports connected to Emulex and QLogic FC HBAs. Tested the same code
with FastLinQ HBAs and confirmed that it is able to process the rport
details as expected. So added the qedf driver in supported list of
drivers in print_fcrports()

Signed-off-by: Milan P. Gandhi <mgandhi@redhat.com>