scsishow: Run fc_rport checks irrespective of scsi_target state
The run_target_checks() function in scsishow.py checks for the fc_rport
state only when the scsi_target is not in STARGET_RUNNING state. But in
some cases it is observed that fc_rport may be BLOCKED even if the
scsi_target is in running state. In such case it would miss to report
any errors on fc_rport. This patch fixes it by running the fc_rport
checks irrespective of scsi_target state.

Signed-off-by: Milan P. Gandhi <mgandhi@redhat.com>