scsishow: Fix an error in checking the host_busy count
The Scsi_Host struct used to have a host_busy counter, it is now removed
from upstream kernel.  The checks performed in run_host_checks() fn in
scsishow.py uses this counter to verify the number of scsi commands pending
on HBA.

While this variable is removed from upstream kernel, it is still being
used in some of the older version of kernels. So, the run_host_checks()
fn first verifies if the Scsi_Host->host_busy count is present, if not
then it calls get_host_busy() to get the count of pending scsi commands.
But there was an error in above check, due to which run_host_checks()
incorrectly assumed that Scsi_Host->host_busy is available, while it
was no longer present. This patch fixes the same.

Signed-off-by: Milan P. Gandhi <mgandhi@redhat.com>