dmshow: check sizes[] list before processing dev logical_block_size details
The automated checks in dmshow.py script verifies the logical_block_size
of each device in dm-linear target. These checks are only for linear
target.  While processing other target types, the sizes[] list remains
uninitialized and accessing it's contents results in crash. So, adding
a check to verify if the list is initialized and then only access it.

Signed-off-by: Milan P. Gandhi <mgandhi@redhat.com>