dmshow: add check for task state TASK_NOLOAD in check
There is a task state TASK_IDLE that is (TASK_UNINTERRUPTIBLE | TASK_NOLOAD).
Since we are only concerned with UN tasks in dmshow check, account
for the possible presence of TASK_NOLOAD as well and skip if found.
Also, add a small counting check to ensure dmshow is in sync with
crash ps output.

Signed-off-by: John Pittman <jpittman@redhat.com>