nvme: adjust pr_cap() for lack of hd_struct
After hd_struct was removed, pr_cap() no longer worked.  Add the
code needed to look in block_device for the sector count and in the
inode if needed as there is a transitory state where the sector count
within the block_device did not exist.  Also, fix spacing in the pr_cap()
function.

Signed-off-by: John Pittman <jpittman@redhat.com>