scsishow: Use starget_to_rport() function to get fc_rport
There are couple of places in progs/scsishow.py program where the same
code is used to retrieve the fc_rport from scsi_target. We also have a
starget_to_rport() function in LinuxDump/scsi module which could be used
to do the same work. So, it's better to utilize the starget_to_rport()
function and avoid code duplication.

Signed-off-by: Milan P. Gandhi <mgandhi@redhat.com>