scsi: Fix the TUR command entry in opcode_table
The print_queue() function retrieves the command opcode from scsi_cmnd
struct and converts it to hex. During this conversion, the TUR command
opcode gets changed from 0x00 to 0x0. But the opcode_table in scsi
package has 0x00 entry for TUR command, so the match for TUR command
fails. Fixing this by changing the TUR command entry in opcode_table
to 0x0.

Signed-off-by: Milan P. Gandhi <mgandhi@redhat.com>