scsi: Load the modules required for processing SCSI host and target details
The scsi package did not verify if the modules required for specific
SCSI adapter are loaded in crash environment. This used to result in
failures in processing the adapter specific verbose details.

Refactoring of this code has also fixed another bug in loading the
scsi_transport_fc modulei which used to fail the fc rport checks.

Signed-off-by: Milan P. Gandhi <mgandhi@redhat.com>