nfsshow: Add helper mutex_in_use()
As of kernel commit 3ca0ff571b09 ("locking/mutex: Rework mutex::owner"),
the struct mutex no longer has a "count" field, but an "owner" field,
which is an atomic_long_t that encodes both the address of the owner and
a set of flags in the low bits.  Add a helper function to determine
whether the mutex is in use that works for both styles of mutex.

Signed-off-by: Scott Mayhew <smayhew@redhat.com>